refactor: restructure project architecture with feature-based organization
