+++ src/config/packageInfo.ts
