+++ src/features/ui/components/index.ts
@@ -1,0 +1,2 @@
+export * from './boxes';
+export * from './tables';
