+++ src/index.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 import { registerAllCommand } from './commands';
-import { createCommander } from './setup';
+import { createCommander } from './config';
 
 const program = createCommander();
