--- src/ui/components/index.ts
@@ -1,1 +1,0 @@
-export * from './boxes';
